backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.24.0
motor==3.5.3
python-jose==3.1.0
passlib[bcrypt]==1.7.4
pydantic==1.10.14
python-multipart==0.0.6
python-dotenv==1.0.0
bcrypt==4.0.1
